moved sampler to separate util file
